+++ app/src/main/java/org/sopt/at/screen/SignInScreen.kt
@@ -35,7 +35,7 @@ import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
 import org.sopt.at.component.BackButton
 import org.sopt.at.component.CommonTextField
-import org.sopt.at.component.NoRippleInteractionSource
+import org.sopt.at.util.NoRippleInteractionSource
 import org.sopt.at.component.PasswordTextField
 import org.sopt.at.ui.theme.TivingTheme
 import org.sopt.at.viewmodel.AuthViewModel
