+++ app/src/main/java/org/sopt/at/viewmodel/AuthViewModel.kt
@@ -5,12 +5,12 @@ import androidx.lifecycle.ViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import org.sopt.at.data.RequestSignInDto
-import org.sopt.at.data.RequestSignUpDto
-import org.sopt.at.data.ResponseNicknameDto
-import org.sopt.at.data.ResponseSignInDto
-import org.sopt.at.data.ResponseSignUpDto
-import org.sopt.at.data.ServicePool
+import org.sopt.at.data.model.RequestSignInDto
+import org.sopt.at.data.model.RequestSignUpDto
+import org.sopt.at.data.model.ResponseNicknameDto
+import org.sopt.at.data.model.ResponseSignInDto
+import org.sopt.at.data.model.ResponseSignUpDto
+import org.sopt.at.data.api.ServicePool
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
